Move calc parser to separate file.
